The budget allocator algorithm as implemented earlier was not consistent with the equations in the dissertation and the implementation had to be corrected.
The system was modified to allow SRT tasks to export the alpha parameter to the allocator task in the form of a 32-bit fixed-point value with 16 fractional bits.

The pbsSRT library, the allocator task, and the pbs_module had to be modified accordingly.

+++ pbs_module/pbs_timing.c
@@ -155,7 +155,7 @@ void sched_period_tick(void)
     struct SRT_timing_struct* SRT_timing_struct_p;
     struct SRT_struct *SRT_struct_p;
 
-    int sp_till_deadline, sp_per_tp;
+    u16 sp_till_deadline, sp_per_tp;
 
     /*Determine the number of instruction retired, energy consumed, and time spent in 
     each mode of operation over the previous reservation period*/
